Bug 1630655 - Ensure PreferencesWriter::Flush actually flushes r=njn
I'm not positive that the intended behavior of PreferencesWriter::Flush
is to ensure that any pending writes to the preferences file are
completed, but that seems like what it ought to be for, and it does
not look water-tight. For one, adding a sync runnable to a thread pool
will not ensure that any previously submitted runnables have completed
by the time it returns. And two, the exchange on sPendingWriteData only
guarantees that the write has started.

This change simply ensures that the write to disk has been completed
before returning.

Differential Revision: https://phabricator.services.mozilla.com/D72016

UltraBlame original commit: b9e4c3d3eca0bfe4ae9d4fb7a7847e66f32b88d3