Bug 1633704 [wpt PR 23087] - [server-timing] Populate values for NavigationTiming, a=testonly
Automatic update from web-platform-tests
[server-timing] Populate values for NavigationTiming

Currently, Navigation Timing's Server Timing values are populated when
the entry is first created, which happens when it is first queried. For
long documents, that can happen before the body is loaded, which is the
point in which the response is set on the NavigationTiming entry's
TimingInfo.
This CL fixes that, by setting the response on the TimingInfo earlier,
at the document's commit time.

Change-Id: I10522342a7d92560167bdb7d33956823e588d9d8

Bug: 1068937
Change-Id: I10522342a7d92560167bdb7d33956823e588d9d8

--

wpt-commits: 0084aa71d4d4935fcbb43bd8122328da9d305098
wpt-pr: 23087

UltraBlame original commit: 779273edaf951bb453327d31c839fa0587bfa70e