Bug 1634621 - Move _fullZoom initialization to resetFields() rather than the browser constructor. r=Gijs
Bug 1633938 re-introduced this field to handle full-zoom on RDM without
having to affect the actual content zoom, see related hacks that went
away in this push:

  https://hg.mozilla.org/integration/autoland/rev/9e6925ea790f994f09c1af3e442aaab1d14388f6

But I added it to the browser element constructor, instead of to
resetFields(), which means that it's not properly reset when the
browsing context changes and thus it can get out of sync.

Move it to resetFields(), so that we properly detect that zoom has
changed. An alternative to this is to only use this._fullZoom when in
RDM instead.

This is where a similar field only used for remote browser used to live
before my Fission zoom changes:

https://searchfox.org/mozilla-central/rev/9642a113737b556750228965bb9cbd3aaa655e22/toolkit/content/widgets/browser-custom-element.js#346

Differential Revision: https://phabricator.services.mozilla.com/D73426

UltraBlame original commit: 626d2d4c0fd627447ebcd96fa1dc4a26f8ea6be3