Bug 1634543 - Fix caret baseline of a block with only empty lines. r=jfkthame
Even with the fix above, we get the caret baseline wrong because the
line is non-empty, but the baseline is all the way to the top.

Maybe an alternative to this would be to move the line-height hack to
the beginning of layout but that sounds a bit more fishy.

Otherwise we should do this more often, I suspect. I added one test that
fails because of nsInlineFrame... Anyhow probably follow-up material?

Depends on D73362

Differential Revision: https://phabricator.services.mozilla.com/D73363

UltraBlame original commit: a38299795286f775856652593dee810cff56a6d7