Bug 1634456 - Use double-raf to wait for caret scrolling to happen.
Before bug 1634153, we were using Dispatch(), so setTimeout(0) was the right
thing to wait for.

In bug 1634153 I used the existing mechanism which works as an early refresh
driver runner, so we need to wait for a refresh driver tick, not an event loop
turn.

Differential Revision: https://phabricator.services.mozilla.com/D73834

UltraBlame original commit: 5f1e56715b5aee0166e7ef154aabb42c71756db2