Bug 1634204 - Allow newer versions of symbols if they are weakly linked. r=froydnj
We currently check that the binaries we ship are not using symbol
versions of system libraries that would not be available on some older
systems. In some cases, however, we may get dependencies on newer symbol
versions in the form of weak symbols, that are checked for at runtime.
This happens with __cxa_thread_atexit_impl when building against a glibc
newer than 2.18, and the supporting code in Rust libstd actually checks
at runtime whether the weak symbol is resolved before using it.

Differential Revision: https://phabricator.services.mozilla.com/D73782

UltraBlame original commit: 71b00ec52717892c4d601971d21ef5ad2ecd5561