Bug 1635664 - Don't capture the output from sdkmanager in ensure_android_packages. r=rstewart
We always print the output when there is no error. In case of error, we
stick the output in the thrown exception, but nothing actually prints
that out. It's simpler to just let the subprocess print its own output
instead of capturing it, so that important error messages are not hidden
in the case of failure.

Differential Revision: https://phabricator.services.mozilla.com/D74004

UltraBlame original commit: b251328e237f57bb2cfe21d167cefa22704bc3b6