Bug 1635001 - P1. Make threading model clearer when dispatching tasks on the controller thread. r=kats
RemoteContentController::PostDelayedTask must be called on the controller thread; and all the 3 implementations are doing is dispatching the task on the current MessageLoop.

The naming and that the method was pure virtual made it a bit confusing as it gave the impression we would dispatch the task on the controller's internal thread, which wasn't the case.

So we make a generic implementation and assert to the documented use.

Differential Revision: https://phabricator.services.mozilla.com/D73828

UltraBlame original commit: 3cb0a05be7fcba9d6a89bbe0208b1aa84f74bab1