Bug 1608771 - Part 4, BigInt<>I64 conversion for inlined calls r=lth,wingo
This is part 4 of implementing the Wasm BigInt<>I64 conversion proposal for inlined Ion to Wasm calls.

This part adds the support for I64/BigInt arguments for inlined calls in the MCallOptimize.cpp part of
the IonBuilder. With this commit, the I64 arguments will work on 64-bit platforms, except where arguments
are required to be spilled to the stack due to the ABI (this case is more complicated to support).

On 32-bit platforms, this commit disables the BigInt/I64 support for the inlined call as it does not work
without further changes.

Differential Revision: https://phabricator.services.mozilla.com/D65237

UltraBlame original commit: f036cb46d18f81e51ed3e7584d2a4eb5b2449447