Bug 1636304. Fix and improve the Mermaid documentation. r=sylvestre
The second example, when copied and pasted, is completely borked since it
is not preceeded by the line that declares what follows to be a Mermaid
diagram. This change fixes that.

It also fixes the prose to read slightly less awkwardly in places.

Finally, this change also adds a link to the Mermaid Live Editor to make
readers aware of this useful tool.

Differential Revision: https://phabricator.services.mozilla.com/D74338

UltraBlame original commit: 65fc0df9b6fa24e16b29620a67b15ebeddbec508