Bug 1635437 - Directly store targetFronts in toolbox redux store. r=ochameau.
We add `targetType` and `isTopLevel` properties on the target front, which
are set from the targetList and are used in the EvaluationContextSelector.

As a result we are able to directly call targetFront.targetType in a few
places instead of calling `getTargetType` again.

Differential Revision: https://phabricator.services.mozilla.com/D73890

UltraBlame original commit: 9192b355750621af88356c3aebc01690dc8b114f