Bug 1634135: Fresh import of irregexp r=mgaudet
This is as good a time as any to pull in some recent upstream changes (many of which I wrote). This patch was auto-generated using import-irregexp.py. The necessary changes to the shim code are in the next patch.

Depends on D73116

Differential Revision: https://phabricator.services.mozilla.com/D73117

UltraBlame original commit: 9dddfd577a3debad921b96bdad5fc24dae5f1093