Bug 1635917 - Correct RegsToPreserve in WasmStubs. r=bbouvier
RegsToPreserve must contain the correct float registers *and* be immune to
PushRegsInMask stripping out the SIMD registers in non-SIMD builds before
it performs register set reduction.  The correct mask to start with is
therefore AllMask, not AllVector128Mask, as the latter is stripped down to
an empty set, not reduced down to the set of doubles.

This bug came in with the SIMD patch and affects non-SIMD builds on x64
as well as all builds on x86.

Differential Revision: https://phabricator.services.mozilla.com/D74825

UltraBlame original commit: 8d060521f7750738e9475283ad6f7da680cf0e7b