Bug 1635261 - Modify the way how we get the top-level principal in AntiTrackingUtils::HasStoragePermissionInParent(). r=dimi,baku
We used to use the topLevelStorageAreaPrincipal and topLevelPrincipal to
get the principal we need to do the storage permission checks in
AntiTrackingUtils::HasStoragePermissionInParent(). But, they don't work
in Fission.

So, we change to use the
AntiTrakcingUtils::GetTopLevelStorageAreaWindowId() and
AntiTrackingUtils::GetTopLevelAntiTrackingWindowID().

Differential Revision: https://phabricator.services.mozilla.com/D74238

UltraBlame original commit: 05ac80df0e76d0c231e278a7951269fc28dcaf29