Bug 1637189: Don't clobber current_position_ when it's non-volatile r=jandem
Unicode case-insensitive backreferences crash on 32-bit x86, but not on 64-bit. The problem is in the code deciding which registers to save before calling CaseInsensitiveCompareUCStrings. We save all the volatile registers, except for a couple that we know we won't need.

After we do that, we modify current_position_, with the implicit expectation that it was saved. That's true on x86-64, but not true on x86-32. The old version of this code used the same approach, but assigned its registers in a different order in the NativeRegExpMacroAssembler constructor, so we got away with it.

The fix is to ensure we always save/restore current_position, even if it's not volatile.

Differential Revision: https://phabricator.services.mozilla.com/D74900

UltraBlame original commit: c289ffcd2510d3121bb4738bb06d4d844a621728