Bug 1607784 - Avoid JS_FreezeObject when creating wasm exports object. r=lth
Profiles indicate that JS_FreezeObject can be expensive for large exports
objects. It looks like properties are re-allocated. It would be better to
just build the exports object correctly the first time. In the attached
test-case to this bug, this commit dropped instantiation time from 160ms
to 70ms.

According to MDN,
```
An object is frozen if and only if it is not extensible, all its properties are non-configurable, and all its data properties (that is, properties which are not accessor properties with getter or setter components) are non-writable.
```

We can accomplish this by changing the JSPROP_FLAG used for each data property
and marking the object as non-extensible at the end.

Differential Revision: https://phabricator.services.mozilla.com/D74968

UltraBlame original commit: ea5b26e729f1c40e30f7898adb757292f0dbbd88