Bug 1636797 - Don't include .pyc, .pyd, or .pyo files in hash in taskgraph r=tomprince,glandium
These files are ALWAYS ignored and are never relevant when hashing the state of a working tree. It would be better overall to not consult the filesystem directly and go through the VCS to ensure we never try to hash any ignored files, but these files seem to be the main stumbling block and the primary cause of bugs like bug 1636797, so this is a fine stopgap in the meantime.

Differential Revision: https://phabricator.services.mozilla.com/D75126

UltraBlame original commit: 465ff2025b7ff9753f7f1ff79d3e0bd56e19b77d