Bug 1636995 - Fix scrolling issue with error messages. r=Honza.
The PageError component wasn't passing the maybeScrollToBottom
prop to the GripMessageBody, which was making the console out-of-autoscroll
since the stacktrace is rendered asynchronously.
Some tests cases are added (throwing directly from an evaluation,
or in a setTimeout), so we cover different paths.

Differential Revision: https://phabricator.services.mozilla.com/D74708

UltraBlame original commit: f1e8a2272911635d8e43d7845db12c129a6a6bf6