Bug 1637868 - Make Baseline free from frame's tls r=wingo,lth
We are going to get rid of the TLS slot in the Frame because we don't
need it all the time. Now it consumes one stack slot in Frame and forces
the Frame to add one additional padding slot for arm64.

In this patch we are removing dependency from Frame::tls in Baseline.
Note that we need to allocate a stack slot for WasmTlsReg value because
baseline can clobber WasmTlsReg and we should be able to load it when we
need it. For example, when we need access to the Instance*.

The followup patch will do the same for the Ion compiler.

Differential Revision: https://phabricator.services.mozilla.com/D66201

UltraBlame original commit: a9970903e180d4e6392b6aa2e27875d35249327c